refactor: 재사용 interface 분리
